Added CalendarReducer
